Move python application client outside of standalone test directory
