+++ tests/application_client/setup.cfg
@@ -1,0 +1,2 @@
+[tool:pytest]
+pythonpath = .
